// src/app/models/receiptHeader.model.ts
export class ReceiptHeader {
      RecieptType: number;
      FileAs: string;
      WhatFor: string; 
      CurrencyId: string;
      Total: number;
      associationId: number;
      EmployeeId: string;
      ThanksLetterId: number;
      Credit4Digit: string;
      CityName?: string;
      CountryCode: string;
      Street: string;
      Street2: string;
      Zip: string;
      fname: string;
      lname: string;
      Titel: string;
      MiddleName?: string;
      Company: string;
      Safix?: string;
      WhatForInThanksLet?: string;
      TotalInLeadCurrent: number;
      CustomizeLine: string;
      CustomerCode: string;
      SendByEmailTo: string;
}