// src/app/models/receiptlines.model.ts
export class Receiptlines {
      ProjectId: number;
      PayTypeId: number;
      Amount: number;
      ValueDate: string;
      CheckNo: number;
      Bank: number;
      BranchNo: string;
      AccountNo: string;
      details?: string;
      DonationTypeId: number;
      AccountId: number;
      AmountInLeadCurrent: number;
}